--- packages/linejs/rate-limit/index.ts
@@ -1,5 +1,0 @@
-/**
- * @module
- * RateLimitter for Safety of LINEJS Client
- */
-export { RateLimitter } from "../client/libs/rate-limitter/index.ts";
