--- packages/linejs/storage/index.ts
@@ -1,13 +1,0 @@
-/**
- * @module
- * The Storage for LINE Client
- */
-
-export {
-	BaseStorage,
-	type Storage,
-} from "../client/libs/storage/base-storage.ts";
-export { MemoryStorage } from "../client/libs/storage/memory-storage.ts";
-export { FileStorage } from "../client/libs/storage/file-storage.ts";
-export { DirStorage } from "../client/libs/storage/dir-storage.ts";
-export { CacheManager } from "../client/libs/storage/cache-manager.ts";
