+++ packages/linejs/utils/thrift/declares.ts
@@ -1,6 +1,6 @@
 import { Buffer } from "node:buffer";
 import { TBinaryProtocol, TCompactProtocol } from "thrift";
-import type { LooseType } from "../../entities/common.ts";
+import type { LooseType } from "../../client/entities/common.ts";
 export const genHeader = {
 	3: (name: string) => {
 		return Buffer.from([
