+++ packages/linejs/utils/typed-event-emitter/index.test.ts
